fix bug so can only delete ads when logged in as proper user

+++ src/main/java/com/codeup/adlister/controllers/DeleteAdServlet.java
@@ -2,6 +2,7 @@ package com.codeup.adlister.controllers;
 
 import com.codeup.adlister.dao.DaoFactory;
 import com.codeup.adlister.models.Ad;
+import com.codeup.adlister.models.User;
 import com.sun.scenario.effect.impl.sw.sse.SSEBlend_SRC_OUTPeer;
 
 import javax.servlet.ServletException;
@@ -20,9 +21,14 @@ public class DeleteAdServlet extends HttpServlet {
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        System.out.println(request.getSession().getAttribute("id"));
         Ad ad = DaoFactory.getAdsDao().getAdById(Integer.parseInt(request.getParameter("id")));
-        DaoFactory.getAdsDao().deleteAd(ad);
+        if (request.getSession().getAttribute("user") == null){
+            response.sendRedirect("/ads");
+            return;
+        }
+        User user = (User) request.getSession().getAttribute("user");
+        if (ad.getUserId() == user.getId())
+            DaoFactory.getAdsDao().deleteAd(ad);
         response.sendRedirect("/ads");
     }
 }
